demo: Extract shared DAM-image-to-URL helper and rework buildOrganization
The logic that resolves a DamImageBlock to an absolute image URL was private to
`buildOrganization`. The Article structured data (added next) needs the same
resolution for its `image` field, so extract it into `damImageToAbsoluteUrl`.

Also split `buildOrganization` into `buildOrganizationNode` (the bare
`Organization` object, for nesting as author/publisher) and the context-wrapped
variant used at the page root, and have both take the content scope -- resolving
the site URL from it -- instead of receiving a pre-resolved URL. The layout
passes the scope accordingly. No change to the emitted output.
