+++ demo/site/src/util/structuredData/buildOrganization.ts
@@ -1,32 +1,15 @@
-import { generateImageUrl } from "@comet/site-nextjs";
-import type { DamImageBlockData, SiteSettingsContentBlockData } from "@src/blocks.generated";
+import type { SiteSettingsContentBlockData } from "@src/blocks.generated";
+import type { ContentScope } from "@src/site-configs";
+import { getSiteConfigForDomain } from "@src/util/getSiteConfigs";
 import type { Organization, WithContext } from "schema-dts";
 
-function toAbsoluteUrl(url: string, siteUrl: string): string {
-    return url.startsWith("http") ? url : new URL(url, siteUrl).toString();
-}
+import { damImageToAbsoluteUrl } from "./damImageToAbsoluteUrl";
 
-function buildLogoUrl(logo: DamImageBlockData, siteUrl: string): string | undefined {
-    const props = logo.block?.props;
+// schema-dts types `Organization` as a union that includes `string`; the node builders only ever produce the object form.
+export type OrganizationNode = Exclude<Organization, string>;
 
-    if (!props) {
-        return undefined;
-    }
-
-    if ("urlTemplate" in props && props.damFile?.image) {
-        const { width, height } = props.damFile.image;
-        const url = generateImageUrl({ src: props.urlTemplate, width }, width / height);
-        return toAbsoluteUrl(url, siteUrl);
-    }
-
-    if (props.damFile?.fileUrl) {
-        return toAbsoluteUrl(props.damFile.fileUrl, siteUrl);
-    }
-
-    return undefined;
-}
-
-export function buildOrganization(content: SiteSettingsContentBlockData, siteUrl: string): WithContext<Organization> | null {
+export function buildOrganizationNode(content: SiteSettingsContentBlockData, scope: ContentScope): OrganizationNode | null {
+    const siteUrl = getSiteConfigForDomain(scope.domain).url;
     const { organization } = content;
 
     const name = organization.name.trim();
@@ -35,7 +18,7 @@ export function buildOrganization(content: SiteSettingsContentBlockData, siteUrl
         return null;
     }
 
-    const logo = buildLogoUrl(organization.logo, siteUrl);
+    const logo = damImageToAbsoluteUrl(organization.logo, siteUrl);
     const sameAs = organization.sameAs.blocks
         .filter((block) => block.visible)
         .map((block) => block.props.url.trim())
@@ -43,7 +26,6 @@ export function buildOrganization(content: SiteSettingsContentBlockData, siteUrl
     const description = organization.description.trim();
 
     return {
-        "@context": "https://schema.org",
         "@type": "Organization",
         name,
         url: organization.url.trim() || siteUrl,
@@ -51,4 +33,14 @@ export function buildOrganization(content: SiteSettingsContentBlockData, siteUrl
         ...(sameAs.length > 0 ? { sameAs } : {}),
         ...(description ? { description } : {}),
     };
+}
+
+export function buildOrganization(content: SiteSettingsContentBlockData, scope: ContentScope): WithContext<Organization> | null {
+    const node = buildOrganizationNode(content, scope);
+
+    if (!node) {
+        return null;
+    }
+
+    return { "@context": "https://schema.org", ...node };
 }
